Crop timeframe reference
Used USDA CDL and a Michigan shapefile to develop code that will count 2023 cropland pixels and rank the top 5 most common crops. 
The .xlsx file is a rough chart of the labeled top 5 crops, peak NDVI months, and average harvest season.
These files are a crop-based reference that can be used to help determine optimal timeframe for satellite image selection.